Chatbot/requirements.txt
python-dotenv
langchain
langchain-google-genai
langchain-community
chromadb
gradio
jira
